Split linter code into separate linter.py file
